requirements.txt
amqp==5.1.1
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-extensions==3.1.5
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
flower==1.0.0
humanize==4.0.0
kombu==5.2.4
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
